Move OverlayDB.cpp/.h from libdevcrypto to libdevcore.
